moved create_collector into pybryt.execution.tracing
